requirements.txt
# Audio Analysis Requirements
librosa>=0.10.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Optional: Essentia for advanced audio analysis
# essentia-tensorflow>=2.1b6.dev1034

# Database and HTTP
psycopg2-binary>=2.9.0
aiohttp>=3.8.0
aiofiles>=0.8.0

# Supabase client
supabase>=1.0.0

# Data processing
pandas>=1.3.0

# Logging and utilities
python-dateutil>=2.8.0